Margin of whole order has been adjusted. Things from RapidResponseDesk.utils.jsx have been moved to respective files.
