+++ src/Tools/RapidResponseDesk/RapidResponseDesk.scss
@@ -11,7 +11,7 @@
     overflow-y: scroll;
 
     & > * {
-      margin: 5px 1px 5px 5px;
+      margin: 0 1px 5px 5px;
       background-color: $whiteColor;
     }
   }
