+++ src/VERSION.js
@@ -1,1 +1,5 @@
-export const VERSION_INFO = { tag: 'v1.11.3', branch: 'v1.11.3', commit: '8e08d872495dc5be592c7b91372a0512788e82a3' }
+export const VERSION_INFO = {
+  tag: 'v1.11.3',
+  branch: 'v1.11.3',
+  commit: '8e08d872495dc5be592c7b91372a0512788e82a3',
+};
